@@ -2,24 +2,19 @@
   "name": "vueleton",
   "version": "1.0.6",
   "description": "Vue components with only core features.",
-  "author": "Gerald <i@gerald.top>",
+  "author": "Gerald <gera2ld@live.com>",
   "license": "ISC",
   "homepage": "https://github.com/gera2ld/vueleton#readme",
   "files": [
     "lib"
   ],
   "scripts": {
-    "clean": "del-cli lib",
-    "dev": "rollup -wc rollup.conf.js",
-    "build:js": "rollup -c rollup.conf.js",
-    "build": "cross-env NODE_ENV=production run-s clean ci build:js",
+    "clean": "gulp clean",
+    "dev": "gulp watch",
+    "build": "gulp build",
     "prepublishOnly": "run-s build",
     "lint": "eslint --ext .js,.vue .",
-    "ci": "run-s lint",
-    "start:docs": "cd docs && nuxt",
-    "build:docs": "cd docs && nuxt generate && touch dist/.nojekyll",
-    "upload:docs": "gh-pages -td docs/dist",
-    "deploy": "run-s build:docs upload:docs"
+    "ci": "run-s lint"
   },
   "keywords": [
     "vue",
@@ -31,26 +26,19 @@
   },
   "peerDependencies": {
     "codemirror": "5.x",
-    "vue": "2.x"
+    "vue": "3.x"
   },
   "devDependencies": {
-    "@gera2ld/plaid": "~2.3.0",
-    "@gera2ld/plaid-common-vue": "~2.2.0",
-    "@gera2ld/plaid-rollup": "~2.3.2",
-    "codemirror": "^5.55.0",
-    "del-cli": "^3.0.1",
-    "eslint": "^7.23.0",
-    "gh-pages": "^3.1.0",
-    "highlight.js": "^10.1.1",
-    "marked": "^1.1.0",
-    "nuxt": "^2.13.3",
-    "raw-loader": "^4.0.1",
-    "rollup-plugin-vue": "^5.1.9",
-    "vue": "^2.6.11",
-    "vue-router": "^3.3.4"
+    "@gera2ld/plaid": "~2.4.1-alpha.4",
+    "@gera2ld/plaid-common-vue": "~2.4.1-alpha.3",
+    "codemirror": "^5.65.0",
+    "del": "^6.0.0",
+    "gulp": "^4.0.2",
+    "gulp-postcss": "^9.0.1",
+    "vue": "^3.2.26"
   },
   "dependencies": {
-    "@babel/runtime": "^7.13.10"
+    "@babel/runtime": "^7.16.5"
   },
   "repository": {
     "type": "git",
