@@ -3,20 +3,22 @@
 </template>
 
 <script>
+import CodeMirror from 'codemirror';
+
 const defaultOptions = {
   lineNumbers: true,
   tabSize: 2,
 };
 
 export default {
+  name: 'vl-code',
   props: ['value', 'options'],
   watch: {
     value: 'update',
     options(options, oldOptions) {
       const { cm } = this;
       if (!cm) return;
-      Object.keys({ ...oldOptions, ...options })
-      .forEach((key) => {
+      Object.keys({ ...oldOptions, ...options }).forEach((key) => {
         if (options[key] !== oldOptions[key]) {
           cm.setOption(key, options[key]);
         }
@@ -26,23 +28,22 @@ export default {
   methods: {
     update(value) {
       if (value === this.cached) return;
-      if (this.cm) this.cm.setValue(this.cached = value == null ? '' : `${value}`);
+      if (this.cm)
+        this.cm.setValue((this.cached = value == null ? '' : `${value}`));
     },
   },
   mounted() {
-    const CodeMirror = require('codemirror'); // eslint-disable-line global-require, import/no-unresolved
-    this.cm = CodeMirror(this.$el, Object.assign({}, defaultOptions, this.options));
+    this.cm = CodeMirror(
+      this.$el,
+      Object.assign({}, defaultOptions, this.options)
+    );
     this.$emit('ready', this.cm);
     this.cm.on('change', (cm) => {
-      this.$emit('input', this.cached = cm.getValue());
+      this.$emit('input', (this.cached = cm.getValue()));
     });
     this.update(this.value);
   },
 };
 </script>
 
-<style>
-.CodeMirror {
-  height: auto;
-}
-</style>
+<style src="./style.css"></style>
